Refactor: Split tests into multiple files based on modules under test
